remove redundant $stdout coercion after test libs
On Thu, Apr 14, 2011 at 6:32 PM, Brian Burns wrote:
> However, I just took my working code and changed my assignment of
> $stdout to the IO object to use $stdout.reopen (using the same
> line you added in your latest version) and immediately the
> deadlock errors reappeared.  So, it would appear that even with
> assigning MiniTest::Unit.output = $stdout using the reopen still
> causes deadlocks, whereas assigning $stdout does not.

Good observation.  This indicates that we have not yet found the
root cause of the problem.  In theory, we should not have to
reassign MiniTest::Unit.output = $stdout at all:

1. In master, minitest grabs a reference to $stdout.
2. In worker, we reopen $stdout into the log file.

So in worker, step 2 should also affect minitest's *reference* to
$stdout.  Explicitly reassigning minitest's reference is rendundant:

 a = 'hello'
 b = a             # (step 1)
 a.replace 'world' # (step 2)
 b = a             # (step 3) redundant! b still points to a

I'm glad you pointed this out.  There really is something else
going on to spoil the works, but I'm not sure what it could be.
Perhaps it's time we asked the experts on ruby-core mailing list?