Fix SIGCHLD handling and test completion.
We need to reap all ready children in SIGCHLD, not just the first, and
should not be reporting completion in the signal handler.

On a fast machine, or where there is a slow hook executed at the
completion of a test run, more than one test child can terminate before
the SIGCHLD handler is invoked, or while it is running.

In that event we will only get another SIGCHLD when a new child
terminates, not to signal that there was more than one current
termination.  We need to loop to collect all terminated children during
each invocation of the handler.

Since we don't know which child terminated, we wait on any terminated
child with NOHANG, until it informs us that there are no more zombies
hanging about.

Doing all the work of finishing the test case, cleaning up, and running
user hooks inside the SIGCHLD handler block was pretty slow.  This could
lead to a big pile-up of children that needed to be cleaned up,
especially if the user hook did something like run another external
process to signal completion.

Moving the heavy work of completion outside the signal handler makes the
whole thing a lot faster, and less likely to bump into the low limit for
per-user processes on Mac OS-X.