replace lock with queue to support SIGCHLD handler
The SIGCHLD handler in tork-master(1) can be triggered at any time, even
in the middle of writing to the standard output stream!  Locking access
to the output stream in normal code (outside the signal handler) would
freeze the program because the signal handler, waiting for the lock to
be released, would never return!

One solution is to maintain a thread-safe queue of outgoing items that
need to be written to the output stream.  Both normal code and the
signal handler can quickly push an outgoing item onto the queue and
proceed with their business.  A separate thread can then have the sole
responsibility of (and access to) continually writing those outgoing
items to the output stream.