tork-runner: percentage lied; print counts instead
The percentage of completion is not accurate because it only measures
how many tests that have started running _so far_ have been completed.
This makes the percentage jump from 100% to lesser values as tests run.

Also drop the `tork-runner:` prefix from progress messages and replace
VT100 escape codes for reprinting in-place with a \r (carriage return).